+++ coders/meta.c
@@ -143,7 +143,7 @@ static MagickBooleanType IsMETA(const unsigned char *magick,const size_t length)
 
 typedef struct _html_code
 {
-  const short
+  const short int
     len;
 
   const char
@@ -225,7 +225,7 @@ static size_t convertHTMLcodes(char *s, const size_t len)
       codes=sizeof(html_codes)/sizeof(html_code);
       for (i=0; i < codes; i++)
       {
-        if (html_codes[i].len <= len)
+        if (html_codes[i].len <= (ssize_t) len)
           if (stringnicmp(s, html_codes[i].code,(size_t) (html_codes[i].len)) == 0)
             {
               (void) memmove(s+1,s+html_codes[i].len,
